Allocator: Use it in more places, pass it to DLLs, and warn about cleanup issues in debug

+++ sources/code/Editor/ImguiEditor/TracingPanel.cpp
@@ -69,7 +69,7 @@ void TracingPanel::RenderContents() {
 		"##CPU Profiling",
 		&ValuesGetter,
 		&session,
-		session.results.size(),
+		static_cast<int>(session.results.size()),
 		0,
 		"Main Thread",
 		FLT_MAX,
